refactor : controller 통합
