+++ src/main/java/cart/product/controller/ProductController.java
@@ -1,41 +1,47 @@
 package cart.product.controller;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
 
 import cart.product.dto.ExceptionResponse;
 import cart.product.dto.ProductRequest;
 import cart.product.dto.ProductResponse;
 import cart.product.service.ProductService;
 
-@RestController
-@RequestMapping("/products")
-public class ProductApiController {
+@Controller
+@RequestMapping("/")
+public class ProductController {
 
     private final ProductService productService;
 
-    public ProductApiController(ProductService productService) {
+    public ProductController(ProductService productService) {
         this.productService = productService;
     }
 
-    @PostMapping
+    @GetMapping(path = {"/", "/admin"})
+    public String displayHome(Model model, HttpServletRequest request) {
+        model.addAttribute("products", productService.findAll());
+        if (request.getRequestURI().equals("/admin")) {
+            return "admin";
+        }
+        return "index";
+    }
+
+    @PostMapping("/products")
+    @ResponseBody
     public ResponseEntity<ProductResponse> createProducts(@Valid @RequestBody ProductRequest productRequest) {
         ProductResponse productResponse = productService.saveProducts(productRequest);
         return ResponseEntity.status(HttpStatus.CREATED).body(productResponse);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("/products/{id}")
+    @ResponseBody
     public ResponseEntity<ProductResponse> updateProducts(@PathVariable Long id,
                                                           @Valid @RequestBody ProductRequest productRequest) {
         ProductResponse productResponse = productService.updateProducts(id, productRequest);
@@ -43,7 +49,8 @@ public class ProductApiController {
         return ResponseEntity.status(HttpStatus.OK).body(productResponse);
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/products/{id}")
+    @ResponseBody
     public ResponseEntity<Object> deleteProducts(@PathVariable Long id) {
         productService.deleteProductsById(id);
